Initial format lambda and container build
Why are these changes being introduced:

* We need to be able to run lambdas as part of the timdex ingest process

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/RDI-48

How does this address that need:

* creates a Dockerfile that imports lambdas
* provides documentation on how to work with the lambdas locally
* includes some initital tests

Document any side effects to this change:

This should allow for multiple lambdas to be built for the timdex
pipline in this single repository and single container but be callable
separately as needed within the pipeline. Each lambda would need to
still be setup, but only one ECR should be required for all of them
in this pipeline.

Co-Authored-By: Helen Bailey <hbailey@mit.edu>