Add transform and starting step
Why these changes are being introduced:
Per the addition of the transmogrifier app, we now have separate
transform and load steps that require different commands. We also
wanted the ability to start the ETL process at any of the three steps
for debugging and testing, since we often don't want to rerun the
harvest and may sometimes not want to rerun the transform either.

How this addresses that need:
* Adds transform section to the format lambda_handler output, with the
  commands and result file information needed for that step
* Renames the ingest step to load to reflect its current functionality
* Adds an input value "starting-step", which can be any of the three
  steps in the harvest-transform-load process
* Updates tests to reflect changes, also adds unittest so they can be
  run easily
* Updates README to reflect changes, including instructions for running
  tests

Side effects of this change:
The input and output requirements have changed, so any existing
resources that use this lambda function or provide input to it
(Eventbridge, Step Functions) will need to be updated.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/RDI-79