Setup initial auto-deploy
Why these changes are being introduced:
This is the first step in setting up the ability for GitHub Actions to
auto-deploy Lambda functions in AWS. It adds a new caller workflow and
it updates the existing Makefile. For now, the caller workflow is using
a shared workflow that is in a feature branch on our shared .github
repository in MITLibraries.

How this addresses that need:
* Create dev-build.yml caller workflow. This calls a shared workflow
that is very similar to the existing dev-ecr-push.yml workflow.
* Use an share GitHub secret in the dev-build.yml workflow for tracking
the correct AWS Account where the Lambda should be deployed
* Use the OIDC GHA role in the dev-build.yml workflow for GitHub
Actions to access the correct AWS Account (this line in the workflow
is generated by the terraform repo that sets up all the infrastructure
for the timdex pipelines)
* Enable `workflow_dispatch` in the new worflow to allow a developer to
manually kick off the deploy-to-Dev1 process in GitHub for testing
purposes.
* Enable push.pull_request trigger inn the new workflow for any PR
issued against the main branch of this repository
* Add additional variables to the Makefile that capture values from
Parameter Store (those values are put into Param Store by the terraform
infrastructure repository)
* Add in the lint commands to the Makefile (since I found these in
Makefiles in other Python repositories)
* Leave in place the local developer specific make commands so a
dev working on this code can manually force a docker push and a lambda
update from their local workstation as-needed. These commands have
been updated slightly to force the connection to ONLY the Dev1 account.
* Add make commands specific to running in a GitHub Actions runner (all
prefixed with `gha-`). These leverage the variables set at the top of
the Makefile in order to dynamically connect to the correct AWS Account,
the correct ECR repository, and the correct Lambda function.

Side effects of this change:
The only side effect is that any PR to main will trigger a re-deploy
of the container to the ECR and the Lambda function will get updated.