Lambda Deploy Automation
Why these changes are being introduced:
These changes enable automated Lambda function builds, uploads, and
deploys in the dev environment. This is linked to opening a PR against
the `main` branch of this repo. The changes here do NOT include any
automation for either the Stage of the Prod environments.

How this addresses that need:
* Update `Makefile` header with output from `terraform apply` on the
associated infrastructure repository:
mitlib-tf-sandbox-timdex-infrastructure
* Create new `dev-build.yml` workflow (generated by `terraform apply`
on the  associated infrastructure repository)

Side effects of this change:
None.