Update makefile with copy/pasted commands
# Subject
Applies the standard makefile commands to the app.  This is mostly just a formatting change

# Why these changes are being introduced:
This commit is part of a larger refactor of how we handle ECR repositories and generating makefile and github workflows.
The work here will likely be removed when stage and prod are deployed, so this is a temporary change, but serves as a proof of concept/stopgap until stage is ready.
The changes here represent a direct copy/paste from the infrastructure repository.

# How this addresses that need:
Update the makefile with our new standard format

# Side effects of this change:
None

# Relevant ticket(s):
NA