Add index-prefix option and complete refactor
Why these changes are being introduced:

The index-prefix option needed to be added so we can specify in the
load step command that the OpenSearch index name should include the
prefix. This is necessary for RDI sources to be searched properly in the
RDI UI separately from other TIMDEX sources.

While adding this, however, it was evident that the format lambda logic
has become complex enough to warrant a significant rewrite with addition
of tests covering all of the logic. Adding our standard python testing
and linting/code quality checks helped to ensure that this all went
smoothly and didn't introduce more problems.

How this addresses that need:
* Restructures the repo to include our usual package folders, with the
  lambda functions in a lambdas package and tests in a tests package.
* Adds pipenv for dependency management and our Python test/lint/code
  quality libraries to the dev dependencies.
* Refactors format.py into format_input.py (format is a standard library
  variable in Python and so not an ideal name for a function) and breaks
  the module down into logical functions.
* Adds some input validation to ensure that invalid input values are
  handled with useful error messages.
* Adds tests covering all functionality in format_input.py.
* Updates Dockerfile to reflect restructured package.
* Updates Makfile to include standard developer commands for testing,
  code checking, and dependency management.
* Adds a setup.cfg file for basic testing and linting configuration.
* Updates README to provide information on the required and optional
  input fields, Makeful usage, and updated examples.
* Adds standard PR template and Github Actions CI workflow.

Side effects of this change:

Any tools that call this lambda (Step Functions, Eventbridge, etc.)
need to use the new input JSON structure. None are currently configured
for automation in AWS so this shouldn't affect any existing
infrastructure.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/RDI-210
* https://mitlibraries.atlassian.net/browse/RDI-73