Update format lambda for new automation changes
Why these changes are being introduced:
With the introduction of Alma records, we need the ability to handle
multiple files of records during each step of the pipeline. There are
also special S3 file handling tasks that need to happen for Alma
records, and we need to check how many files were exported after the
extract phase. We will also need the ability to handle deleted records
in the very near future, which this commit does not add but does set us
up for.

How this addresses that need:
* Adds a commands module with three output command generators, one for
  each phase (extract, transform, load).
* Updates the format lambda to call one of the above generators,
  determined by an input field value.
* Adds an alma_prep module that does S3 file handling (extracting zipped
  files to the TIMDEX bucket) as a pre-transform step for Alma files
  only.
* Refactors the input validation into a new config module.
* Adds helper methods and an error class to support TIMDEX-specific
  logic that is applicable to multiple pipeline steps, such as file
  naming.
* Adds/updates tests and fixtures for all new/changed functionality.
* Updates the README to reflect new input/output expectations and new
  required ENV variables.
* Minor updates to gitignore, setup.cfg, Makefile

Side effects of this change:
The step function JSON will need to be updated as well, and the old step
function will not work as it has different input/output expectations.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-112
* https://mitlibraries.atlassian.net/browse/TIMX-113