Update format input response for no files
Why these changes are being introduced:
Now that we are sending notifications within the step function, we need
certain data to be present in the input to each step of the state
machine (run date, source, etc.) so we can include that information in
the notification message. While we were previously including that in the
happy path format_input workflow, we were not including it in either the
success or failure case of no files present after extraction. This
updates those responses to also include the basic input data about the
step function run.

How this addresses that need:
* In format_input.py, adds no-files-present success and failure state
  data to the initial result data set and returns it all, instead of
  just returning the success/failure data.
* Updates tests to reflect expected lambda function responses.

Side effects of this change:
SNS notifications should now work when no files are present after the extraction step, both in the success case (there were no updated files to harvest) and the failure case (we expected files to be present but there weren't any).

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-127