Update project to python 3.11 and new linting
Why these changes are being introduced:

We are systematically moving through our python applications and updating the python version, linting configurations, and
some other general scaffolding.

With this project likely touched for some upcoming projects, it was decided to update in advance of that work.

How this addresses that need:

  * bumps python version to 3.11
  * adds ruff and pyproject.toml approach to linting
  * code style and conventions updated to match new linting rules

Side effects of this change:

The python version was updated from 3.10 to 3.11.  Tests have completed fine and testing will be performed in Dev1, but
majoy python versions always come with some risk of unintended changes.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-894