Establish feature flag branching for parquet work
Why these changes are being introduced:
* We will be using a feature flag approach while modifying
TIMDEX Pipeline Lambdas to generate the required commands
for the Transform step to output files to a parquet dataset
and for the Load step to read records a parquet dataset
and index records into TIMDEX.

We are doing so with the goal of full backwards compatibility until
that refactor is complete.

How this addresses that need:
* Utilizes a new optional env var 'ETL_VERSION' to provide an ETL version
for code to branch from.
* Format Input Lambda function handler is updated to provide an ETL version
to the functions that generate commands for the Transform and Load step.
* Add branching logic and submethods to the functions that generate commands
for the Transform and Load step.
* All feature flag logic branching is noted by comments, suitable for
removal when the development work is complete.

Side effects of this change:
* TIMDEX Pipeline Lambdas remains fully backwards compatible, either via
the absence of env var 'ETL_VERSION' or if the value equals '1'.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-412