Temporary tests for feature flagging
Why these changes are being introduced:
* We do not need to heavily test the feature flagging logic,
knowing it will be removed after development work.  But a bit
of testing can ensure that it works as expected while termporarily
in a deployed state.

How this addresses that need:
* Adds new test file tests/test_temporary_feature_flagging.py that
is noted to be removed after development work is complete
* Test branching logic in Format Input handler using 'ETL_VERSION' env var
* Test config function for retrieving 'ETL_VERSION' env var

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-412