Explicitly update TDA for make update
Why these changes are being introduced:

The TDA library is a Github dependency in the Pipefile, but the
normal `pipenv update --dev` does not pickup versions bumps in the
TDA library.  However, an explicit `pipenv update <TDA_URL>` will.

How this addresses that need:
* Adds an explicit TDA update in the `make update` command

Side effects of this change:
* TDA is updated to v2.0 (current)

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-496