Add run-timestamp to transform commands
Why these changes are being introduced:

Similar to when run-id was added as an allowed payload attribute, that
is then passed around to various command generation, so the same is needed
for run-timestamp for generating transform commands (for Transmogrifier).

How this addresses that need:
* "run-timestamp" is an allowed input payload attribute
* if included, passed to transform commands
* if absent, a run-timestamp is minted by the lambda for all transform
commands generated

The net effect is the lambda will provide the *same* run-timestamp for
all Transmogrifier commands it prepares, which ensures all writes for
the run get the same timestamp.

The only variation is whether the StepFunction passes the timestamp,
or the lambda mints it; both are supported.

Side effects of this change:
* All Transmogrifier commands will now include a --run-timestamp CLI
argument

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-509