Update location for ETL read and writes
Why these changes are being introduced:

We have decided to move where ETL records are stored in the
TIMDEX ETL dataset.  Instead of living at the root of the
/dataset directory, they will now organize under /dataset/data/records.

The reasoning goes beyond the scope of this commit, but this application
is one of the key places in the pipeline where the location to write
(Transmogrifier) and read (TIM) records is defined.

How this addresses that need:

Building on previous commits that moved configurations into a Config
class, this creates a new property Config.s3_etl_records_data_location
which can be used throughout the application when the root of the ETL
parquet dataset is needed.

Side effects of this change:
* Transmogrifier will begin writing to /dataset/data/records instead
of /dataset
* TIM will read from /dataset/data/records instead of /dataset

Note: this change can be made without any other immediate
modifications to the ETL pipeline.  There is current work applying a
new metadata layer in the dataset, but from an ETL point of view,
the write and read locations are defined explicitly by this application
for a single ETL run.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-521