Point to dataset root
Why these changes are being introduced:

After recent updates in TDA, any library that instantiates TIMDEXDataset
should, again, point back to the *root* of the dataset.  Once initialized,
the library knows for data it might be /data/records or for metadata
it might be /metadata.

How this addresses that need:
* Update config property to remove /data/records
* Update tests to remove /data/records

Side effects of this change:
* Expecting updated version of TDA for any library
consuming the output of this lambda

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-531