Use AWS SAM for local testing
Why these changes are being introduced:

Previously, local testing was performed by building a Docker container
and issuing HTTP requests against the container that simulated an
invocation (to be specific, this was not a true HTTP invocation of the lambda
but instead a convenience wrapper in the Docker image to field "regular"
invocations via HTTP requests).  While functional, it does not provide a lot
of the affordances and conveniences of AWS SAM, including:

- explicit HTTP versus regular/direction invocation
- handling of AWS credentials for the running container
- simulation of resources

Related, while the repository is named "timdex-pipeline-lambdas" plural,
it effectively continues to be a single lambda.  It is proposed that we
consider keeping it as a single lambda, but explore introducing more
structured business logic branching in the lambda entrypoint.  It is anticipated
that the pipeline lambda could be helpful for things beyond just CLI command
formatting, at which time attention to entrypoint handling and how we test
will get more important.

How this addresses that need:

Adds AWS SAM scaffolding to the repository, while also updating the
README to guide users towards using this approach.

The `ping.py` module was also removed, which was a placeholder for another
distinct lambda definition.

Side effects of this change:
* Local testing is performed via AWS SAM, which is present in other more
recent AWS Lambda repositories we manage.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/USE-86