New source mitlibwebsite extract command
Why these changes are being introduced:

With the addition of the new source 'mitlibwebsite', we need the
pipeline lambdas to parse the StepFunction input payload and
generate an extract/harvest CLI command.

How this addresses that need:
* 'mitlibwebsite' added as a source in config
* followed established structure for extract/harveste commands

Note: while adding support for 'mitlibwebsite' it's starting to
feel like some refactoring might be in order.  Out of scope for
this commit which is tested and confirmed to work.

Side effects of this change:
* mitlibwebsite now a valid source for extract/harvest commands

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/USE-86