Refactor handling of input payload and output response
Why these changes are being introduced:

Each time a change is required in this lambda, one can feel some of
brittleness that has collected around input handling, argument passing,
and logic branching for extract, transform, and load command generation.

The change that prompted this refactor was related to how and when
we set `success` or `failure` nodes in the response object, where it
was fairly difficult in the code to see how and when they are set.

Ignoring whether or not we ultimately rework how `success` and `failure`
are set as possible nodes in the output property, centralizing the form
of the input and output responses could assist in that work.

How this addresses that need:

The refactor work here is to create two new dataclass objects:
- InputPayload
- ResultPayload

By having typed properties, at a glance we can determine what kind
of properties we can expect on the input payload and what we plan
to return in our response.

Furthermore, by having objects to pass around, this simplifies the
signatures of a lot of helpers that were passing around source, run_date,
run_type, etc., in numerous places.  Now we generally pass around an
`input_payload` object and a `result` object, with functions modifying
those.

Very little changes beyond this.

Side effects of this change:
* No side effects in behavior.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/USE-279
  * while not directly related, it's the ticket that prompted this work