Temporarily restore success and failure response properties
Why these changes are being introduced:

Until changes can be made to the StepFunction that would lean into a controlled
use of `next-step`, we need to keep `success` and `failure` as properties in the
response for early exits.  Primarily, this allows daily harvests with no records
to exit without error, which it is not.

How this addresses that need:
* Restores `success` and `failure` as allowed response properties
* Sets `result.success` and/or `result.failure` as needed for early exits
* Comments are applied indicating these are temporary measures

Side effects of this change:
* StepFunction will not error out for daily harvests that have zero records
* Once StepFunction is updated to utilize `next-step`, we'll want to fully
remove `success` and `failure` from the lambda response

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/USE-279